--- js/utils/check-string-length.js
@@ -1,4 +1,0 @@
-/*eslint-disable*/
-function checkStringLength (string, length) {
-  return string.length <= length;
-}
